refactor: simplify build and packaging of modal
