--- packages/components/modal/.babelrc
@@ -1,10 +1,0 @@
-{
-    "plugins": [
-        ["module-resolver", {
-            "root": ["."],
-            "alias": {
-                "@lit-element-bootstrap/button": "@lit-element-bootstrap/button/src"
-            }
-        }]
-    ]
-}
