+++ packages/components/modal/bs-modal-header.js
@@ -1,6 +1,6 @@
 
 import { LitElement, html } from 'lit-element';
-import { BsModalHeaderCss } from './bs-modal-header.css.js';
+import { BsModalHeaderCss } from './css/bs-modal-header.css.js';
 import { BsContentRebootCss } from '@lit-element-bootstrap/content';
 
 import '@lit-element-bootstrap/button/bs-close-button.js';
