+++ packages/components/modal/css/bs-modal-footer.css.js
