+++ packages/components/modal/css/index.js
@@ -1,8 +1,4 @@
-export { BsModal } from './bs-modal.js';
 export { BsModalCss } from './bs-modal.css.js';
-export { BsModalBody } from './bs-modal-body.js';
 export { BsModalBodyCss } from './bs-modal-body.css.js';
-export { BsModalFooter } from './bs-modal-footer.js';
 export { BsModalFooterCss } from './bs-modal-footer.css.js';
-export { BsModalHeader } from './bs-modal-header.js';
 export { BsModalHeaderCss } from './bs-modal-header.css.js';
