+++ packages/components/modal/index.js
@@ -1,0 +1,5 @@
+export * from './css'
+export { BsModal } from './bs-modal.js';
+export { BsModalBody } from './bs-modal-body.js';
+export { BsModalFooter } from './bs-modal-footer.js';
+export { BsModalHeader } from './bs-modal-header.js';
