+++ packages/components/modal/package.json
@@ -2,8 +2,8 @@
   "name": "@lit-element-bootstrap/modal",
   "version": "1.0.0-alpha.7",
   "description": "Bootstrap v4.1.3 modal component in LitElement",
-  "main": "src/index.js",
-  "module": "src/index.js",
+  "main": "index.js",
+  "module": "index.js",
   "type": "module",
   "author": "Nikolas Christou",
   "homepage": "https://github.com/nik-christou/lit-element-bootstrap",
@@ -12,8 +12,23 @@
     "url": "https://github.com/nik-christou/lit-element-bootstrap"
   },
   "license": "MIT",
+  "files": [
+    "index.js",
+    "index.d.ts",
+    "global.d.ts",
+    "bs-*.js",
+    "bs-*.d.ts",
+    "css/*.js",
+    "css/*.d.ts",
+    "custom-elements.json"
+  ],
   "scripts": {
-    "pika:build": "pika build",
+    "prepack": "yarn build",
+    "build": "npm run wca-analyze:json",
+    "lit-analyze": "lit-analyzer bs-*.js css/*.js --failFast",
+    "wca-analyze:json": "wca analyze . --outFile custom-elements.json",
+    "lint:prettier": "prettier \"./**/*.js\" --tab-width 4 --list-different",
+    "format:prettier": "prettier \"./**/*.js\" --tab-width 4 --write",
     "test": "karma start",
     "test:watch": "karma start --auto-watch=true --single-run=false",
     "test:update-snapshots": "karma start --update-snapshots",
@@ -29,31 +44,14 @@
     "lit-html": "^1.1.2"
   },
   "devDependencies": {
-    "@pika/pack": "^0.5.0",
-    "@pika/plugin-build-web": "^0.8.1",
-    "@pika/plugin-bundle-web": "^0.8.1",
-    "@pika/plugin-standard-pkg": "^0.8.1",
     "@open-wc/testing": "^2.4.1",
     "@open-wc/testing-karma": "^3.2.6",
     "@open-wc/testing-karma-bs": "^1.3.6",
-    "babel-plugin-module-resolver": "^3.2.0",
+    "@open-wc/prettier-config": "^0.1.12",
+    "eslint": "^6.6.0",
+    "lit-analyzer": "^1.1.9",
     "karma": "^4.4.1",
-    "sinon": "^7.5.0"
-  },
-  "@pika/pack": {
-    "pipeline": [
-      [
-        "@pika/plugin-standard-pkg"
-      ],
-      [
-        "@pika/plugin-build-web"
-      ],
-      [
-        "@pika/plugin-bundle-web",
-        {
-          "entrypoint": "unpkg"
-        }
-      ]
-    ]
+    "sinon": "^7.5.0",
+    "web-component-analyzer": "^0.1.21"
   }
 }
